Move NodeVisitor, TraverserFactory and NodeTraveser under the PHPParser namespace
