+++ .gitignore
@@ -7,7 +7,7 @@
 /fixtures/set004/scoper.inc.php
 /fixtures/*/vendor
 /vendor/
-/vendor-bin/box/vendor/
+/vendor-bin/*/vendor/
 /.travis/
 !/.travis/.gitkeep
 !/.travis/secrets.tar.enc
