+++ src/PhpParser/NodeVisitor/Resolver/FullyQualifiedNameResolver.php
@@ -12,12 +12,12 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Humbug\PhpScoper\NodeVisitor\Resolver;
+namespace Humbug\PhpScoper\PhpParser\NodeVisitor\Resolver;
 
-use Humbug\PhpScoper\NodeVisitor\AppendParentNode;
-use Humbug\PhpScoper\NodeVisitor\Collection\NamespaceStmtCollection;
-use Humbug\PhpScoper\NodeVisitor\Collection\UseStmtCollection;
-use Humbug\PhpScoper\NodeVisitor\NameStmtPrefixer;
+use Humbug\PhpScoper\PhpParser\NodeVisitor\AppendParentNode;
+use Humbug\PhpScoper\PhpParser\NodeVisitor\Collection\NamespaceStmtCollection;
+use Humbug\PhpScoper\PhpParser\NodeVisitor\Collection\UseStmtCollection;
+use Humbug\PhpScoper\PhpParser\NodeVisitor\NameStmtPrefixer;
 use PhpParser\Node\Expr\ConstFetch;
 use PhpParser\Node\Expr\FuncCall;
 use PhpParser\Node\Name;
