+++ src/Scoper/PhpScoper.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 
 namespace Humbug\PhpScoper\Scoper;
 
+use Humbug\PhpScoper\PhpParser\TraverserFactory;
 use Humbug\PhpScoper\Scoper;
 use PhpParser\Error as PhpParserError;
 use PhpParser\Parser;
