+++ vendor-bin/covers-validator/bin/covers-validator
@@ -1,0 +1,1 @@
+../vendor/ockcyp/covers-validator/covers-validator
