+++ vendor-bin/covers-validator/bin/phpunit
@@ -1,0 +1,1 @@
+../vendor/phpunit/phpunit/phpunit
