+++ vendor-bin/covers-validator/composer.json
@@ -1,0 +1,10 @@
+{
+    "require-dev": {
+        "symfony/console": "^3.3",
+        "ockcyp/covers-validator": "^0.6"
+    },
+    "config": {
+        "bin-dir": "bin",
+        "sort-packages": true
+    }
+}
